+++ CHANGELOG.md
@@ -11,6 +11,7 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 - #844 Validation for BPN to Notification API (Create / Edit), Fixed pagination
 - #726 Added @Preauthorize annotation to dashboard controller
 - #837 Added digital twin type to data provisioning workflow to be able to lookup shells created by trace-x
+- #783 Validation that receiver of a notification must not be same as sender.
 - #859 Enable autocomplete API to filter for given assets
 
 ### Changed
@@ -27,7 +28,9 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 - #778 update EDC from 0.5.3 to 0.7.0
 
 ### Removed
--
+
+- #602 digitalTwinType instead of semanticId. DigitalTwinType causes problems in release 24.05
+
 ## [10.8.4 - 17.04.2024]
 
 ### Added